from float to doubles

+++ examples/scale/main.cpp
@@ -50,7 +50,7 @@ auto make_column()
 
    auto indicator_color = get_theme().indicator_color;
 
-   float const button_scale = 1.0/4;
+   double const button_scale = 1.0/4;
    sprite power_button = sprite{ "power_180x632.png", 158*button_scale, button_scale };
    sprite phase_button = sprite{ "phase_180x632.png", 158*button_scale, button_scale };
    sprite mail_button = sprite{ "mail_180x632.png", 158*button_scale, button_scale };
